--- src/main/java/org/cloud/sonic/agent/common/maps/MiniCapMap.java
@@ -1,13 +1,0 @@
-package org.cloud.sonic.agent.common.maps;
-
-import javax.websocket.Session;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-public class MiniCapMap {
-    private static Map<Session, Thread> miniCapMap = new ConcurrentHashMap<>();
-
-    public static Map<Session, Thread> getMap() {
-        return miniCapMap;
-    }
-}
